[backend] review api 수정
