+++ backend/routes/apiRoutes.js
@@ -5,6 +5,7 @@ const productRoutes = require('./productRoutes');
 const categoryRoutes = require('./categoryRoutes');
 const userRoutes = require('./userRoutes');
 const orderRoutes = require('./orderRoutes');
+const reviewRoutes = require('./reviewRoutes');
 const bannerRoutes = require('./bannerRoutes');
 
 const jwt = require('jsonwebtoken');
@@ -42,6 +43,7 @@ app.use('/products', productRoutes);
 app.use('/categories', categoryRoutes);
 app.use('/users', userRoutes);
 app.use('/orders', orderRoutes);
+app.use('/review', reviewRoutes);
 app.use('/banners', bannerRoutes);
 
 module.exports = app;
